Viikko 6 tehtavat 9-11

+++ app/helpers/application_helper.rb
@@ -3,7 +3,11 @@ module ApplicationHelper
   def edit_and_destroy_buttons(item)
     unless current_user.nil?
       edit = link_to('Edit', url_for([:edit, item]), class:"btn-sm btn-primary")
-      del = link_to('Delete', item, method: :delete, data: {confirm: 'Are you sure?' }, class:"btn-sm btn-danger")
+      if current_user.admin
+        del = link_to('Delete', item, method: :delete, data: {confirm: 'Are you sure?' }, class:"btn-sm btn-danger")
+      else
+        del = ""
+      end
       raw("#{edit} #{del}")
     end
   end
